Mean Shift is a hierarchical clustering algoritm.
Mean Shift is a hierarchical clustering algorithm. In contrast to supervised machine learning,
clustering attempts to group data without labeling, the latter is very useful in the context of
Business Intelligence for study which cluster belongs to a specific set of business's indicators.
As opposed to the KMeans, when using MeanShift, you don't need to know the numbers of clusters (categories) beforehand.